serverless-application/requirements.txt
aws-cdk.core==1.129.0
aws_cdk.aws_dynamodb==1.129.0
aws-cdk.aws_s3_notifications==1.129.0
aws-cdk.aws-lambda==1.129.0
aws-cdk.aws-s3==1.129.0